+++ crates/base/core/src/code/marker/mod.rs
@@ -8,11 +8,15 @@
 
 mod _reexport; // SYMLINK from /src/code/marker/_reexport_core.rs
 
+mod prim; // Prim, PrimFitPtr
 mod type_marker; // zero-cost generic type markers
 
 crate::structural_mods! { // _mods, _reexports
     _mods {
-        pub use super::type_marker::*;
+        pub use super::{
+            prim::*,
+            type_marker::*,
+        };
     }
     _reexports {
         pub use super::_reexport::*;